added few database models and crud

+++ backend/tamalesHr/src/main/java/com/sjsu/cmpe272/tamales/tamalesHr/repository/EmployeeRepository.java
@@ -3,7 +3,4 @@ package com.sjsu.cmpe272.tamales.tamalesHr.repository;
 import com.sjsu.cmpe272.tamales.tamalesHr.model.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeRepository extends JpaRepository<Employee, Long> {
-
-
-}
+public interface EmployeeRepository extends JpaRepository<Employee, Integer> {}
